kafka_producer/requirements.txt
kafka-python==2.0.2
faker==22.5.1
